Поправки на основе замечаний
Разделил на дополнительные модули, вынес в папку mocks
